requirements.txt
pyOpenSSL==19.0.0
requests==2.21.0